read nested categoricals
