+++ polars/polars-core/Cargo.toml
@@ -177,7 +177,7 @@ package = "arrow2"
 git = "https://github.com/ritchie46/arrow2"
 # rev = "98e49133b2e56e51e30335830485b3cf768eb5a2"
 # path = "../../../arrow2"
-branch = "apply_validity"
+branch = "polars"
 # version = "0.12"
 default-features = false
 features = [
