+++ polars/polars-core/src/chunked_array/mod.rs
@@ -38,6 +38,7 @@ pub mod strings;
 pub mod temporal;
 mod trusted_len;
 pub mod upstream_traits;
+mod from;
 
 use polars_arrow::prelude::*;
 
@@ -46,6 +47,7 @@ use crate::chunked_array::categorical::RevMapping;
 use crate::series::IsSorted;
 use crate::utils::CustomIterTools;
 use bitflags::bitflags;
+use polars_arrow::kernels::concatenate::concatenate_owned_unchecked;
 use std::mem;
 
 #[cfg(not(feature = "dtype-categorical"))]
@@ -454,84 +456,6 @@ where
         } else {
             slice(self)
         }
-    }
-}
-
-impl<T> ChunkedArray<T>
-where
-    T: PolarsDataType,
-{
-    /// Create a new ChunkedArray from existing chunks.
-    pub fn from_chunks(name: &str, chunks: Vec<ArrayRef>) -> Self {
-        // prevent List<Null> if the inner list type is known.
-        let datatype = if matches!(T::get_dtype(), DataType::List(_)) {
-            if let Some(arr) = chunks.get(0) {
-                arr.data_type().into()
-            } else {
-                T::get_dtype()
-            }
-        } else {
-            T::get_dtype()
-        };
-        let field = Arc::new(Field::new(name, datatype));
-        let mut out = ChunkedArray {
-            field,
-            chunks,
-            phantom: PhantomData,
-            categorical_map: None,
-            bit_settings: Default::default(),
-            length: 0,
-        };
-        out.compute_len();
-        out
-    }
-}
-
-// A hack to save compiler bloat for null arrays
-impl Int32Chunked {
-    pub(crate) fn new_null(name: &str, len: usize) -> Self {
-        let arr = arrow::array::new_null_array(ArrowDataType::Null, len);
-        let field = Arc::new(Field::new(name, DataType::Null));
-        let chunks = vec![arr as ArrayRef];
-        let mut out = ChunkedArray {
-            field,
-            chunks,
-            phantom: PhantomData,
-            categorical_map: None,
-            bit_settings: Default::default(),
-            length: 0,
-        };
-        out.compute_len();
-        out
-    }
-}
-
-impl<T> ChunkedArray<T>
-where
-    T: PolarsNumericType,
-{
-    /// Create a new ChunkedArray by taking ownership of the Vec. This operation is zero copy.
-    pub fn from_vec(name: &str, v: Vec<T::Native>) -> Self {
-        let arr = to_array::<T>(v, None);
-        Self::from_chunks(name, vec![arr])
-    }
-
-    /// Nullify values in slice with an existing null bitmap
-    pub fn new_from_owned_with_null_bitmap(
-        name: &str,
-        values: Vec<T::Native>,
-        buffer: Option<Bitmap>,
-    ) -> Self {
-        let arr = to_array::<T>(values, buffer);
-        let mut out = ChunkedArray {
-            field: Arc::new(Field::new(name, T::get_dtype())),
-            chunks: vec![arr],
-            phantom: PhantomData,
-            categorical_map: None,
-            ..Default::default()
-        };
-        out.compute_len();
-        out
     }
 }
